Reorder signup logic
