--- endpoints/instantiations/auth_/signup.py
@@ -1,3 +1,0 @@
-from utilities.namespace import create_namespace, AppNamespaces
-
-namespace_signup = create_namespace(AppNamespaces.AUTH)
