+++ tests/helper_scripts/helper_classes/RequestValidator.py
@@ -341,7 +341,7 @@ class RequestValidator:
         expected_json_content: Optional[dict] = None,
         expected_response_status: HTTPStatus = HTTPStatus.OK,
     ):
-        mock = mocker.patch("middleware.primary_resource_logic.signup.send_signup_link")
+        mock = mocker.patch("endpoints.instantiations.auth_.signup.middleware.send_signup_link")
         self.post(
             endpoint="/api/auth/signup",
             json={"email": email, "password": password},
